Split code into more files
