+++ server/transaction_commit_informer.go
@@ -22,11 +22,6 @@ import (
 	"sync"
 	"time"
 
-	"github.com/cloudwan/gohan/db"
-	"github.com/cloudwan/gohan/db/options"
-	"github.com/cloudwan/gohan/db/transaction"
-	"github.com/cloudwan/gohan/metrics"
-	"github.com/cloudwan/gohan/schema"
 	gohan_sync "github.com/cloudwan/gohan/sync"
 	"github.com/cloudwan/gohan/util"
 )
@@ -108,188 +103,4 @@ func buildSyncValue(id int64) string {
 	}
 
 	return string(data)
-}
-
-//DbSyncWrapper wraps db.DB so it logs events in database on every transaction.
-type DbSyncWrapper struct {
-	db db.DB
-}
-
-func NewDbSyncWrapper(db db.DB) db.DB {
-	return &DbSyncWrapper{db}
-}
-
-// BeginTx wraps transaction object with sync
-func (sw *DbSyncWrapper) BeginTx(options ...transaction.Option) (transaction.Transaction, error) {
-	tx, err := sw.db.BeginTx(options...)
-	if err != nil {
-		return nil, err
-	}
-	return syncTransactionWrap(tx), nil
-}
-
-func (sw *DbSyncWrapper) Connect(dbType string, conn string, maxOpenConn int) error {
-	return sw.db.Connect(dbType, conn, maxOpenConn)
-}
-
-func (sw *DbSyncWrapper) Close() {
-	sw.db.Close()
-}
-
-func (sw *DbSyncWrapper) RegisterTable(s *schema.Schema, cascade, migrate bool) error {
-	return sw.db.RegisterTable(s, cascade, migrate)
-}
-
-func (sw *DbSyncWrapper) DropTable(s *schema.Schema) error {
-	return sw.db.DropTable(s)
-}
-
-func (sw *DbSyncWrapper) Options() options.Options {
-	return sw.db.Options()
-}
-
-type transactionEventLogger struct {
-	transaction.Transaction
-	eventLogged bool
-	lastEventId int64
-}
-
-func syncTransactionWrap(tx transaction.Transaction) *transactionEventLogger {
-	return &transactionEventLogger{tx, false, badEventId}
-}
-
-func (tl *transactionEventLogger) logEvent(ctx context.Context, eventType string, resource *schema.Resource, version int64) error {
-	schemaManager := schema.GetManager()
-	eventSchema, ok := schemaManager.Schema("event")
-	if !ok {
-		return fmt.Errorf("event schema not found")
-	}
-
-	if resource.Schema().Metadata["nosync"] == true {
-		log.Debug("skipping event logging for schema: %s", resource.Schema().ID)
-		return nil
-	}
-
-	body, err := resource.JSONString()
-
-	syncPlain := false
-	syncPlainRaw, ok := resource.Schema().Metadata["sync_plain"]
-	if ok {
-		syncPlainBool, ok := syncPlainRaw.(bool)
-		if ok {
-			syncPlain = syncPlainBool
-		}
-	}
-
-	syncProperty := ""
-	syncPropertyRaw, ok := resource.Schema().Metadata["sync_property"]
-	if ok {
-		syncPropertyStr, ok := syncPropertyRaw.(string)
-		if ok {
-			syncProperty = syncPropertyStr
-		}
-	}
-
-	if err != nil {
-		return fmt.Errorf("Error during event resource deserialisation: %s", err.Error())
-	}
-	eventResource := schema.NewResource(eventSchema, map[string]interface{}{
-		"type":          eventType,
-		"path":          resource.Path(),
-		"version":       version,
-		"body":          body,
-		"sync_plain":    syncPlain,
-		"sync_property": syncProperty,
-		"timestamp":     int64(time.Now().Unix()),
-	})
-	tl.eventLogged = true
-
-	result, err := tl.Transaction.Create(ctx, eventResource)
-	if err != nil {
-		return err
-	}
-
-	tl.lastEventId, err = result.LastInsertId()
-	return err
-}
-
-func (tl *transactionEventLogger) Create(ctx context.Context, resource *schema.Resource) (transaction.Result, error) {
-	result, err := tl.Transaction.Create(ctx, resource)
-	if err != nil {
-		return nil, err
-	}
-	if err = tl.logEvent(ctx, "create", resource, 1); err != nil {
-		return nil, err
-	}
-
-	return result, nil
-}
-
-func (tl *transactionEventLogger) Update(ctx context.Context, resource *schema.Resource) error {
-	err := tl.Transaction.Update(ctx, resource)
-	if err != nil {
-		return err
-	}
-	if !resource.Schema().StateVersioning() {
-		return tl.logEvent(ctx, "update", resource, 0)
-	}
-	state, err := tl.StateFetch(ctx, resource.Schema(), transaction.IDFilter(resource.ID()))
-	if err != nil {
-		return err
-	}
-	return tl.logEvent(ctx, "update", resource, state.ConfigVersion)
-}
-
-func (tl *transactionEventLogger) Resync(ctx context.Context, resource *schema.Resource) error {
-	if !resource.Schema().StateVersioning() {
-		return tl.logEvent(context.Background(), "update", resource, 0)
-	}
-	state, err := tl.StateFetch(ctx, resource.Schema(), transaction.IDFilter(resource.ID()))
-	if err != nil {
-		return err
-	}
-	return tl.logEvent(context.Background(), "update", resource, state.ConfigVersion)
-}
-
-func (tl *transactionEventLogger) Delete(ctx context.Context, s *schema.Schema, resourceID interface{}) error {
-	resource, err := tl.Fetch(ctx, s, transaction.IDFilter(resourceID), nil)
-	if err != nil {
-		return err
-	}
-	configVersion := int64(0)
-	if resource.Schema().StateVersioning() {
-		state, err := tl.StateFetch(ctx, s, transaction.IDFilter(resourceID))
-		if err != nil {
-			return err
-		}
-		configVersion = state.ConfigVersion + 1
-	}
-	err = tl.Transaction.Delete(ctx, s, resourceID)
-	if err != nil {
-		return err
-	}
-	return tl.logEvent(ctx, "delete", resource, configVersion)
-}
-
-func (tl *transactionEventLogger) Commit() error {
-	err := tl.Transaction.Commit()
-	if err != nil {
-		return err
-	}
-
-	tl.triggerSyncWriter()
-	return nil
-}
-
-func (tl *transactionEventLogger) triggerSyncWriter() {
-	if !tl.eventLogged {
-		return
-	}
-
-	if tl.lastEventId == badEventId {
-		panic("logic error, lastEventId not set")
-	}
-
-	transactionCommitted <- tl.lastEventId
-	metrics.UpdateCounter(1, "event_logger.notified")
 }
